encode chat flags as a byte instead of int to fix vanilla 1.3 compatibility

+++ src/main/java/org/spout/vanilla/protocol/codec/LocaleViewDistanceCodec.java
@@ -55,7 +55,7 @@ public class LocaleViewDistanceCodec extends MessageCodec<LocalViewDistanceMessa
 		ChannelBuffer buffer = ChannelBuffers.dynamicBuffer();
 		ChannelBufferUtils.writeString(buffer, message.getLocale());
 		buffer.writeByte(message.getViewDistance());
-		buffer.writeInt(message.getChatFlags());
+		buffer.writeByte(message.getChatFlags());
 		buffer.writeByte(message.getDifficulty());
 		return buffer;
 	}
